Adds `interface` kind in order to parse Doxygen-generated C# output.
When Doxygen parses C#, it is possible to generate an object with
`kind='interface'`. While Doxygen can then convert this output type into
generated documentation, `exhale` doesn't recognize this kind of output object
and the generated source files don't include any RST for the interfaces. As a
result, building the resulting documentation source tree with `sphinx` (or with
`breathe`) results in missing pages and missing cross-links.